move Story\Memory trait under Relay
